refactor(web): extract agent scaffold + Md generators
Move agent-creation helpers into src/web/agent-scaffold.ts:

- ensureAgentHooks (idempotent settings.json hooks migration)
- writeAgentSettingsFromProfile (profile-driven permission writer)
- scaffoldAgentDir (initial directory + file layout)
- generateClaudeMd, generateSoulMd, generateSkillMd
  (LLM-backed CLAUDE.md / SOUL.md / SKILL.md scaffolders)

web.ts: 4564 -> 4371 lines. No behavior change.
